requirements.txt
streamlit
openai
python-dotenv
PyPDF2
tiktoken